Updating event timing and comments
George, I consolidated/cleaned up a bunch of the comments, but tried not to meaningfully change your intent in the comments. Overall this is a very illustrative and informative example! 

I did add the event timing of "C" since the second one was "with food". I think this is appropriate. The IG schematron makes it look like only 1 timing is allowed, but it's embedded with a "such that" so the [0..1] binding doesn't prohibit additional effectiveTime elements according to my reading.